move AuthenticationInfo from dispaly package to user
