+++ zeppelin-display/src/test/scala/org/apache/zeppelin/display/angular/AbstractAngularElemTest.scala
@@ -19,8 +19,9 @@ package org.apache.zeppelin.display.angular
 import java.io.{ByteArrayOutputStream, PrintStream}
 import java.util
 
-import org.apache.zeppelin.display.{AuthenticationInfo, AngularObject, AngularObjectRegistry, GUI}
+import org.apache.zeppelin.display.{AngularObject, AngularObjectRegistry, GUI}
 import org.apache.zeppelin.interpreter._
+import org.apache.zeppelin.user.AuthenticationInfo
 import org.scalatest.concurrent.Eventually
 import org.scalatest.{BeforeAndAfter, BeforeAndAfterEach, FlatSpec, Matchers}
